Import FirebaseSharedSwift and add test helpers
